# cmd/api/docs/swagger.yaml
definitions:
  domain.Category:
    properties:
      category:
        type: string
      id:
        type: integer
    type: object
...
  models.AddCart:
    properties:
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
...
  models.RemoveFromCart:
    properties:
      id:
        type: integer
      product_id:
        type: integer
    type: object
  models.SetNewName:
    properties:
      current:
        type: string
      new:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Show Times is an E-commerce platform to purchase Watch
  title: Go + Gin E-Commerce API Show Times
  version: 1.0.0
paths:
...
  /admin/categories:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all available categories.
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Retrieved categories successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Fields provided are in the wrong format'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Could not retrieve categories'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get all categories
      tags:
      - Category Management
  /admin/category:
    post:
      consumes:
      - application/json
      description: Adds a new category based on the provided details.
      parameters:
      - description: Category details to add
        in: body
        name: AddCategory
        required: true
        schema:
          $ref: '#/definitions/domain.Category'
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Category added successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Fields are provided in the wrong format'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Could not add the category'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Add a new category
      tags:
      - Category Management
  /admin/category/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a category based on the provided category ID.
      parameters:
      - description: Category ID to delete
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Category deleted successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Fields are not provided in the correct format'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Could not delete the category'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Delete a category
      tags:
      - Category Management
  /admin/category/update:
    patch:
      consumes:
      - application/json
      description: Updates the name of an existing category based on the provided
        details.
      parameters:
      - description: Current and new category name
        in: body
        name: UpdateCategory
        required: true
        schema:
          $ref: '#/definitions/models.SetNewName'
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Category updated successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Fields provided are in the wrong format'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Could not update the category'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Update category name
      tags:
      - Category Management
...
  /user/cart:
    patch:
      consumes:
      - application/json
      description: Updates the quantity of a product in the user's cart based on the
        provided details.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product details to update quantity
        in: body
        name: UpdateCart
        required: true
        schema:
          $ref: '#/definitions/models.AddCart'
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Quantity updated successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Cannot update quantity or fields are in the wrong
            format'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Update failed'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Update product quantity in cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Adds an item to the user's cart based on the provided details.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Item details to add to the cart
        in: body
        name: AddCart
        required: true
        schema:
          $ref: '#/definitions/models.AddCart'
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Item added to cart successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Fields provided in the wrong format'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Cannot add item to cart'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Add an item to the cart
      tags:
      - Cart
  /user/cart/list:
    get:
      consumes:
      - application/json
      description: Retrieves the list of items in the user's cart based on the user
        ID.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Retrieved cart items successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Cannot list products'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Could not retrieve the cart list'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Retrieve cart items
      tags:
      - Cart
  /user/cart/remove:
    delete:
      consumes:
      - application/json
      description: Removes a product from the user's cart based on the provided details.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product details to remove from cart
        in: body
        name: RemoveFromCart
        required: true
        schema:
          $ref: '#/definitions/models.RemoveFromCart'
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Product removed from cart successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Cannot remove product or fields are in the wrong
            format'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Removing from cart failed'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Remove product from cart
      tags:
      - Cart
...
securityDefinitions:
  BearerTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"